64 bits port: updated data type to avoid warnings.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@46812 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/rout_obj.c
@@ -170,8 +170,8 @@ rt_public void rout_obj_call_procedure_dynamic (BODY_INDEX body_id, EIF_ARG_UNIO
 	void* body_addr;
 	void* patt_addr;
 	struct item* it;
-	int count;
-	int i = 0;
+	size_t count;
+	size_t i = 0;
 
 	count = strlen ((char*)types);
 
@@ -255,8 +255,8 @@ rt_public void rout_obj_call_function_dynamic (BODY_INDEX body_id, EIF_ARG_UNION
 	void* patt_addr;
 	void* resp;
 	struct item* it;
-	int count;
-	int i = 0;
+	size_t count;
+	size_t i = 0;
 
 	count = strlen ((char*)types);
 	resp = *(EIF_REFERENCE *) res;
